refact: 오케스트레이터 도입 및 레거시 정리
