+++ recommendation-service/src/main/java/com/happymapleday/recommendation/service/impl/OptimizationServiceImpl.java
@@ -5,7 +5,7 @@ import com.happymapleday.recommendation.dto.response.CharacterRecommendation;
 import com.happymapleday.recommendation.dto.response.OptimizationSummary;
 import com.happymapleday.recommendation.service.OptimizationService;
 import com.happymapleday.recommendation.service.calculator.OptimizationSummaryCalculator;
-import com.happymapleday.recommendation.service.limiter.CrystalLimitManager;
+import com.happymapleday.recommendation.service.orchestrator.OptimizationOrchestrator;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -15,7 +15,7 @@ import java.util.List;
 @RequiredArgsConstructor
 public class OptimizationServiceImpl implements OptimizationService {
     
-    private final CrystalLimitManager crystalLimitManager;
+    private final OptimizationOrchestrator optimizationOrchestrator;
     private final OptimizationSummaryCalculator optimizationSummaryCalculator;
     
     // 각 캐릭터별 제약 조건을 고려한 최적화된 추천 결과 생성
@@ -26,7 +26,7 @@ public class OptimizationServiceImpl implements OptimizationService {
         }
 
         // 전체 최적화 로직 사용
-        return crystalLimitManager.optimizeGlobalRecommendations(characterBossSelections);
+        return optimizationOrchestrator.optimizeGlobalRecommendations(characterBossSelections);
     }
     
     // 최적화 요약 생성
